export_data_from_gcs_to_box/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-storage
google-cloud-secret-manager
boxsdk[jwt]
regex